MB-9343 wait for seq no persistence
When doing a rebalance we want to avoid to possibility of
massive data loss immediately after a VBucket move.
This can happen if we don’t persist everything to disk before
starting the next VBucket move.

As a result we should wait for persistence of a majority of
the items before pausing indexing. The rebalancer can make sure
items up to a certain sequence number are persisted by using
the “Wait For Persistence” command.

This command responds with either a timeout, which means the
items have not yet been persisted, or a success meaning that
all items up to the given sequence number are persisted.

Change-Id: I3118566e56de7ad7e9be061f7726d6f377975b77